+++ sentry-android-core/src/main/java/io/sentry/android/core/ManifestMetadataReader.java
@@ -127,8 +127,7 @@ final class ManifestMetadataReader {
   static final String ENABLE_AUTO_TRACE_ID_GENERATION =
       "io.sentry.traces.enable-auto-id-generation";
 
-  static final String AUTO_TRANSACTION_DEADLINE_TIMEOUT_MILLIS =
-      "io.sentry.traces.auto-transaction-deadline-timeout-millis";
+  static final String DEADLINE_TIMEOUT = "io.sentry.traces.deadline-timeout";
 
   static final String FEEDBACK_NAME_REQUIRED = "io.sentry.feedback.is-name-required";
 
@@ -449,12 +448,8 @@ final class ManifestMetadataReader {
                 ENABLE_AUTO_TRACE_ID_GENERATION,
                 options.isEnableAutoTraceIdGeneration()));
 
-        options.setAutoTransactionDeadlineTimeoutMillis(
-            readLong(
-                metadata,
-                logger,
-                AUTO_TRANSACTION_DEADLINE_TIMEOUT_MILLIS,
-                options.getAutoTransactionDeadlineTimeoutMillis()));
+        options.setDeadlineTimeout(
+            readLong(metadata, logger, DEADLINE_TIMEOUT, options.getDeadlineTimeout()));
 
         if (options.getSessionReplay().getSessionSampleRate() == null) {
           final double sessionSampleRate =