+++ sentry-android-core/src/main/java/io/sentry/android/core/internal/gestures/SentryGestureListener.java
@@ -251,7 +251,7 @@ public final class SentryGestureListener implements GestureDetector.OnGestureLis
     transactionOptions.setWaitForChildren(true);
 
     // Set deadline timeout based on configured option
-    final long deadlineTimeoutMillis = options.getAutoTransactionDeadlineTimeoutMillis();
+    final long deadlineTimeoutMillis = options.getDeadlineTimeout();
     // No deadline when zero or negative value is set
     transactionOptions.setDeadlineTimeout(
         deadlineTimeoutMillis <= 0 ? null : deadlineTimeoutMillis);